function should still return a double I guess? Only not in the case of !

+++ 6hst/factorial_calculator.cpp
@@ -145,7 +145,7 @@ double primary()
 
 //Edit Dirk
 //
-int factorial()
+double factorial()
 {
 	double left = primary();
 	Token t = ts.get();
